+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# CHANGELOG
+
+- Now the stream returned by `StreamUseCase` and `ProgressUseCase` is not closed when they receive an error
+- `FlutterConsoleLogOutput` has been added to avoid conflicting flutter prints with logger prints
+- `DateTimeSerializerPlugin` has been added
+
 ## 0.0.1
 
 * TODO: Describe initial release.